removed print

+++ src/Config/Simulation_Config.py
@@ -8,12 +8,12 @@ from config.Aero_Config import Aero
 
 
 # Main differential equation 
-def derivative(state: np.array, t: int, rocket: RocketConfig, motor: Motor) -> np.ndarray:
+def derivative(state: np.array, t: float, rocket: RocketConfig, motor: Motor) -> np.ndarray:
     """State space equation to be integrated numericaly. 
 
     Args:
         state (np.array): State Vector [altitude - m, velocity - m/s, mass -kg]
-        t (int): Time of current step in integratoin - seconds
+        t (float): Time of current step in integratoin - seconds
         rocket (RocketConfig): Rocketfig class containing perameters/method of rocket
         motor (Motor): Motor class conatin peramters/methods of the motor
